+++ .circleci/config.yml
@@ -1303,8 +1303,6 @@ jobs:
             JEST_JUNIT_OUTPUT_DIR: ./reports/
       - store_test_results:
           path: libs/ui/reports/
-      - store_artifacts:
-          path: libs/ui/src/reports/__image_snapshots__/__diff_output__/
 
   # @votingworks/usb-drive
   test-libs-usb-drive:
