+++ libs/printing/src/__image_snapshots__/write-in-image-report-1.png
