+++ libs/printing/src/__image_snapshots__/write-in-image-report-empty-1.png
