+++ libs/printing/src/__image_snapshots__/write-in-image-report-primary-1.png
